+++ resources/js/Components/BuiltIn/AuthenticationCardLogo.vue
