+++ resources/js/Components/BuiltIn/Banner.vue
