+++ resources/js/Components/BuiltIn/SectionTitle.vue
