+++ resources/js/Components/BuiltIn/Welcome.vue
@@ -1,5 +1,5 @@
 <script setup>
-import ApplicationLogo from "@/Components/ApplicationLogo.vue";
+import ApplicationLogo from "@/Components/BuiltIn/ApplicationLogo.vue";
 </script>
 
 <template>
