+++ resources/js/Components/Input.vue
@@ -1,0 +1,11 @@
+<script setup>
+
+</script>
+
+<template>
+  $END$
+</template>
+
+<style scoped>
+
+</style>
